+++ api/ring-camera.ts
@@ -9,7 +9,7 @@ import {
   SnapshotTimestamp
 } from './ring-types'
 import { clientApi, RingRestClient } from './rest-client'
-import { BehaviorSubject, Subject } from 'rxjs'
+import { BehaviorSubject, interval, Subject } from 'rxjs'
 import {
   distinctUntilChanged,
   filter,
@@ -17,12 +17,14 @@ import {
   publishReplay,
   refCount,
   share,
-  take
+  take,
+  takeUntil
 } from 'rxjs/operators'
 import { createSocket } from 'dgram'
 import { bindToPort, getPublicIp, reservePorts, SrtpOptions } from './rtp-utils'
 import { delay, logError, logInfo } from './util'
 import { FfmpegOptions, SipSession } from './sip-session'
+import { SipOptions } from './sip-call'
 
 const snapshotRefreshDelay = 500,
   maxSnapshotRefreshSeconds = 30,
@@ -77,7 +79,7 @@ export class RingCamera {
   )
   onInHomeDoorbellStatus = this.onData.pipe(
     map(({ settings: { chime_settings } }: CameraData) => {
-      return Boolean(chime_settings && chime_settings.enable)
+      return Boolean(chime_settings?.enable)
     }),
     distinctUntilChanged()
   )
@@ -86,7 +88,25 @@ export class RingCamera {
     private initialData: CameraData,
     public isDoorbot: boolean,
     private restClient: RingRestClient
-  ) {}
+  ) {
+    if (!initialData.subscribed) {
+      this.subscribeToDingEvents().catch(e => {
+        logError(
+          'Failed to subscribe ' + initialData.description + ' to ding events'
+        )
+        logError(e)
+      })
+    }
+
+    if (!initialData.subscribed_motions) {
+      this.subscribeToMotionEvents().catch(e => {
+        logError(
+          'Failed to subscribe ' + initialData.description + ' to motion events'
+        )
+        logError(e)
+      })
+    }
+  }
 
   updateData(update: CameraData) {
     this.onData.next(update)
@@ -202,34 +222,59 @@ export class RingCamera {
   }
 
   startVideoOnDemand() {
-    return this.restClient.request({
+    return this.restClient.request<ActiveDing | ''>({
       method: 'POST',
-      url: this.doorbotUrl('vod')
+      url: this.doorbotUrl('live_view') // Ring app uses vod for battery cams, but doesn't appear to be necessary
     })
   }
 
+  private pollForActiveDing() {
+    // try every second until a new ding is received
+    interval(1000)
+      .pipe(takeUntil(this.onNewDing))
+      .subscribe(() => {
+        this.onRequestActiveDings.next()
+      })
+  }
+
+  private expiredDingIds: string[] = []
   async getSipConnectionDetails() {
-    const vodPromise = this.onNewDing
-      .pipe(
-        filter(x => x.kind === 'on_demand'),
-        take(1)
-      )
-      .toPromise()
-    await this.startVideoOnDemand()
-    this.onRequestActiveDings.next()
+    const vodPromise = this.onNewDing.pipe(take(1)).toPromise(),
+      videoOnDemandDing = await this.startVideoOnDemand()
+
+    if (
+      videoOnDemandDing &&
+      'sip_from' in videoOnDemandDing &&
+      !this.expiredDingIds.includes(videoOnDemandDing.id_str)
+    ) {
+      // wired cams return a ding from live_view so we don't need to wait
+      return videoOnDemandDing
+    }
+
+    // battery cams return '' from live_view so we need to request active dings and wait
+    this.pollForActiveDing()
     return vodPromise
   }
 
+  private removeDingById(idToRemove: string) {
+    const allActiveDings = this.activeDings,
+      otherDings = allActiveDings.filter(ding => ding.id_str !== idToRemove)
+
+    this.onActiveDings.next(otherDings)
+  }
+
   processActiveDing(ding: ActiveDing) {
-    const activeDings = this.activeDings
+    const activeDings = this.activeDings,
+      dingId = ding.id_str
 
     this.onNewDing.next(ding)
-    this.onActiveDings.next(activeDings.concat([ding]))
+    this.onActiveDings.next(
+      activeDings.filter(d => d.id_str !== dingId).concat([ding])
+    )
 
     setTimeout(() => {
-      const allActiveDings = this.activeDings,
-        otherDings = allActiveDings.filter(oldDing => oldDing !== ding)
-      this.onActiveDings.next(otherDings)
+      this.removeDingById(ding.id_str)
+      this.expiredDingIds = this.expiredDingIds.filter(id => id !== dingId)
     }, 65 * 1000) // dings last ~1 minute
   }
 
@@ -345,23 +390,31 @@ export class RingCamera {
     return this.lastSnapshotPromise
   }
 
-  sipUsedDingIds: string[] = []
-
-  async getSipOptions() {
+  async getSipOptions(): Promise<SipOptions> {
     const activeDings = this.onActiveDings.getValue(),
       existingDing = activeDings
+        .filter(ding => !this.expiredDingIds.includes(ding.id_str))
         .slice()
-        .reverse()
-        .find(x => !this.sipUsedDingIds.includes(x.id_str)),
-      targetDing = existingDing || (await this.getSipConnectionDetails())
+        .reverse()[0],
+      ding = existingDing || (await this.getSipConnectionDetails())
 
-    this.sipUsedDingIds.push(targetDing.id_str)
+    if (this.expiredDingIds.includes(ding.id_str)) {
+      logInfo('Waiting for a new live stream to start...')
+      await delay(500)
+      return this.getSipOptions()
+    }
 
     return {
-      to: targetDing.sip_to,
-      from: targetDing.sip_from,
-      dingId: targetDing.id_str
+      to: ding.sip_to,
+      from: ding.sip_from,
+      dingId: ding.id_str
     }
+  }
+
+  getUpdatedSipOptions(expiredDingId: string) {
+    // Got a 480 from sip session, which means it's no longer active
+    this.expiredDingIds.push(expiredDingId)
+    return this.getSipOptions()
   }
 
   async createSipSession(
@@ -395,13 +448,12 @@ export class RingCamera {
       }
 
     return new SipSession(
-      {
-        ...sipOptions,
-        tlsPort
-      },
+      sipOptions,
       rtpOptions,
       videoSocket,
-      audioSocket
+      audioSocket,
+      tlsPort,
+      this
     )
   }
 
@@ -418,6 +470,34 @@ export class RingCamera {
     const sipSession = await this.createSipSession()
     await sipSession.start(ffmpegOptions)
     return sipSession
+  }
+
+  subscribeToDingEvents() {
+    return this.restClient.request({
+      method: 'POST',
+      url: this.doorbotUrl('subscribe')
+    })
+  }
+
+  unsubscribeFromDingEvents() {
+    return this.restClient.request({
+      method: 'POST',
+      url: this.doorbotUrl('unsubscribe')
+    })
+  }
+
+  subscribeToMotionEvents() {
+    return this.restClient.request({
+      method: 'POST',
+      url: this.doorbotUrl('motions_subscribe')
+    })
+  }
+
+  unsubscribeFromMotionEvents() {
+    return this.restClient.request({
+      method: 'POST',
+      url: this.doorbotUrl('motions_unsubscribe')
+    })
   }
 }